+++ packages/dart_frog_cli/lib/src/commands/dev/dev.dart
@@ -4,6 +4,7 @@ import 'package:dart_frog_cli/src/command.dart';
 import 'package:dart_frog_cli/src/commands/commands.dart';
 import 'package:dart_frog_cli/src/commands/dev/templates/dart_frog_dev_server_bundle.dart';
 import 'package:dart_frog_cli/src/dev_server_runner/dev_server_runner.dart';
+import 'package:dart_frog_cli/src/runtime_compatibility.dart' as runtime_compat;
 import 'package:mason/mason.dart';
 
 /// Typedef for [DevServerRunner.new].
@@ -24,7 +25,10 @@ class DevCommand extends DartFrogCommand {
     super.logger,
     GeneratorBuilder? generator,
     DevServerRunnerBuilder? devServerRunnerBuilder,
-  })  : _generator = generator ?? MasonGenerator.fromBundle,
+    runtime_compat.RuntimeCompatibilityCallback? ensureRuntimeCompatibility,
+  })  : _ensureRuntimeCompatibility = ensureRuntimeCompatibility ??
+            runtime_compat.ensureRuntimeCompatibility,
+        _generator = generator ?? MasonGenerator.fromBundle,
         _devServerRunnerBuilder =
             devServerRunnerBuilder ?? DevServerRunner.new {
     argParser
@@ -46,6 +50,7 @@ class DevCommand extends DartFrogCommand {
 
   final GeneratorBuilder _generator;
   final DevServerRunnerBuilder _devServerRunnerBuilder;
+  final runtime_compat.RuntimeCompatibilityCallback _ensureRuntimeCompatibility;
 
   @override
   final String description = 'Run a local development server.';
@@ -55,6 +60,8 @@ class DevCommand extends DartFrogCommand {
 
   @override
   Future<int> run() async {
+    _ensureRuntimeCompatibility(cwd);
+
     final port = io.Platform.environment['PORT'] ?? results['port'] as String;
     final dartVmServicePort = (results['dart-vm-service-port'] as String?) ??
         _defaultDartVmServicePort;
